+++ assets/js/charitable.js
@@ -968,15 +968,15 @@ CHARITABLE = window.CHARITABLE || {};
                 if ( $form.length ) {
                     $form.append( '<input type="hidden" id="charitable-submit-button-value" />' );
 
-                    $('body').one( 'submit', $form, function( event ) {
-                        var name = document.activeElement.name,
-                            value = document.activeElement.value;
+                    $form.find( '[type=submit]' ).on( 'click', function( event ) {
+                        var name = event.currentTarget.name,
+                            value = event.currentTarget.value;
 
                         $form.find( '#charitable-submit-button-value' )
                             .attr( 'name', name )
                             .attr( 'value', value );
 
-                        $form.find( 'input[type=submit], button[type=submit]' )
+                        $form.find( '[type=submit]' )
                             .attr( 'disabled', 'disabled' );
 
                         return true;