+++ includes/users/class-charitable-user.php
@@ -259,6 +259,9 @@ if ( ! class_exists( 'Charitable_User' ) ) :
 				return $verified;
 			}
 
+			/* Delete the time that the verification email was sent. */
+			delete_user_meta( $this->user->ID, '_charitable_email_verification_email_send_time' );
+
 			/* Check for an existing donor account. */
 			$donor_id = $this->get_donor_id();