+++ templates/account/email-not-verified.php
@@ -16,8 +16,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
+$redirect = array_key_exists( 'redirect_url', $_GET ) ? $_GET['redirect_url'] : false;
+$link     = charitable_get_email_verification_link( get_user_by( 'login', $_GET['login'] ), $redirect, true );
+
 ?>
 <p><?php _e( 'We were unable to verify your email address.', 'charitable' ) ?></p>
 <?php if ( isset( $_GET['login'] ) ) : ?>
-    <p><a href="<?php echo esc_url( charitable_get_email_verification_link( get_user_by( 'login', $_GET['login'] ) ) ) ?>"><?php _e( 'Resend verification email', 'charitable' ) ?></a></p>
+    <p><a href="<?php echo esc_url_raw( $link ) ?>"><?php _e( 'Resend verification email', 'charitable' ) ?></a></p>
 <?php endif ?>